refactor: update domain in manifest to match the home-assistant brands repo naming
