+++ check_all.py
@@ -28,7 +28,7 @@ def main():
     all_passed = True
 
     # Check if we're in the right directory
-    if not Path("custom_components/areas").exists():
+    if not Path("custom_components/custom_areas").exists():
         print("❌ Not in the correct directory. Please run from the project root.")
         return 1
 
@@ -45,13 +45,13 @@ def main():
         all_passed = False
 
     # Run linting
-    if not run_command("black --check --diff custom_components/areas/", "Black formatting check"):
+    if not run_command("black --check --diff custom_components/custom_areas/", "Black formatting check"):
         all_passed = False
 
-    if not run_command("isort --check-only --diff custom_components/areas/", "Import sorting check"):
+    if not run_command("isort --check-only --diff custom_components/custom_areas/", "Import sorting check"):
         all_passed = False
 
-    if not run_command("flake8 custom_components/areas/", "Flake8 linting"):
+    if not run_command("flake8 custom_components/custom_areas/", "Flake8 linting"):
         all_passed = False
 
     # Run pre-commit (if available)
