+++ custom_components/custom_areas/__init__.py
