+++ custom_components/custom_areas/const.py
