+++ custom_components/custom_areas/manifest.json
@@ -1,5 +1,5 @@
 {
-  "domain": "areas",
+  "domain": "custom_areas",
   "name": "Custom Areas Integration",
   "codeowners": ["@DefinitelyADev"],
   "config_flow": true,
