+++ custom_components/custom_areas/sensor.py
