+++ custom_components/custom_areas/strings.json
