+++ custom_components/custom_areas/translations/en.json
