+++ run_mypy.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 # Script to run mypy from the custom_components directory to avoid module name conflicts
-cd custom_components && mypy areas/
+cd custom_components && mypy custom_areas/
